Code simplifications reported by IntelliJ IDEA.
git-svn-id: https://svn.apache.org/repos/asf/roller/trunk@1546773 13f79535-47bb-0310-9956-ffa450edef68

+++ app/src/main/java/org/apache/roller/weblogger/business/jpa/JPAUserManagerImpl.java
@@ -429,7 +429,7 @@ public class JPAUserManagerImpl implements UserManager {
         Query q = strategy.getNamedQuery("User.getCountEnabledDistinct");
         q.setParameter(1, Boolean.TRUE);
         List results = q.getResultList();
-        ret =((Long)results.get(0)).longValue(); 
+        ret =((Long)results.get(0));
 
         return ret;
     }
