+++ app/src/main/java/org/apache/roller/weblogger/webservices/adminprotocol/sdk/WeblogEntry.java
@@ -128,7 +128,7 @@ public class WeblogEntry extends Entry {
         // created
         Element createdElement = e.getChild(Tags.DATE_CREATED, Service.NAMESPACE);
         if (createdElement != null) {
-            setDateCreated(new Date(Long.valueOf(createdElement.getText()).longValue()));
+            setDateCreated(new Date(Long.valueOf(createdElement.getText())));
         }
         
         // APP entries URL
